team12/backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
dj-database-url>=2.1
gunicorn>=21.0
